nfp_net_main: Verify that board.state=15 before attempting the RTSYM lookup
On systems where the host OS boots up faster than the NFP, there
exists a race condition where the NFP is still initializing EMU memories
when the nfp_net.ko drivers is attempting to probe for the control
BAR symbol in the RTSYM table.

To eliminate this race condition, we now check for the HWInfo 'board.state'
variable, which is the value of '15' if and only if the NFP's ARM
initialization has completed.

Signed-off-by: Jason S. McMullan <jason.mcmullan@netronome.com>