NFP Service Processor: Update to support NSP ABI specification v0.3
This obsoletes the earlier 'ARM SP' support, and uses the 'nfp.sp'
resource tag instead of the 'arm.sp' tag.

The v0.3 specifaction adds a 32-bit command code option field, a CPP ID
and CPP address of a buffer (for firmware transfers or DMA) and can
return a 32-bit result in addition to the 8-bit error code.

Signed-off-by: Jason S. McMullan <jason.mcmullan@netronome.com>