nfp.ko: add 'nfp_reset' module parameter
The 'nfp_reset' module parameter on nfp.ko has a slightly different
behaviour than on nfp_net.ko. On nfp.ko, if nfp_reset=1, the module
will _always_ force a nfp_soft_reset() when loaded, regardless of
the whether a firmware load is requested or not.

If a firmware load has occurred, nfp.ko will _always_ attempt a
nfp_soft_reset(), regardless of the state of the nfp_reset= module
paramater.

Signed-off-by: Jason S. McMullan <jason.mcmullan@netronome.com>