nfp_net_common: improve setting device mac
netdev->dev_addr is stored in network order and NFP
expects registers to be written with the value in
little endian.  Also lowest two bytes need to be
put into two highest on NFP.

All this is a bit strange so lets put this piece of
code into a separate function.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>