nfp_netvf: stop using devm_* functions for ctrl_bar
We are using the "managed" API for iomaps but we call iounmap
explicitly anyway which makes no sense.  It's a 50/50 call on
whether to remove the unmaps or stop using devm API.  I'm going
with removal as this is the only usage of "managed" API in the
netdev part of our drivers.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Acked-by: Jason S. McMullan <jason.mcmullan@netronome.com>
Reviewed-by: Rolf Neugebauer <rolf.neugebauer@netronome.com>