nfp_net_common: add usage counting of VXLAN ports
VXLAN tunnels can be established for both IPv4 and IPv6 which
means the same port may be added and removed multiple times
and we need to track the use count.  Otherwise if two tunnels
are added for the same port and one of them gets removed we
will remove the port even though the second tunnel still exists.

Fixes: 1ee3c5600ccf ("nfp_net: add RX encapsulation offload")
Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Jason S. McMullan <jason.mcmullan@netronome.com>