nfp_net_common: split buffer allocation into two functions
Split buffer allocation into the allocation+mapping part and the part
which puts the buffer on the rings.

Since the function which puts the buffer on the ring does the kicking
we need to stash the number of buffers allocated since last kick in
the ring structure and kick only once every BATCH times.  This is
very similar to what we do for .xmit_more.  Note that we have to
split a 32 bit index field to make sure ring still fits into a cache
line.  This should be fine unless we plan to have more than 65k rings.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Rolf Neugebauer <rolf.neugebauer@netronome.com>