nfp_net_common: drop packets on RX buffer allocation failures
The canonical way to handle RX buffer management is to allocate
a buffer before a received packet is taken off the ring.  This
way if allocation fails we can put the buffer back on the ring
effectively discarding the packet.

Thanks to reuse of the packets we can guarantee that ring will
never be empty.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Rolf Neugebauer <rolf.neugebauer@netronome.com>