Makefile: fix error message on make clean
Make clean prints following message:
  cat: .revision: No such file or directory
It's caused by the fact that clean target is not executed in
the same path as others.  Whenever our Makefile sees no ".git"
directory it assumes that it should use .revision file to
find out the version.  Check if the .revision file is present
to avoid such warnings.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Jason S. McMullan <jason.mcmullan@netronome.com>
Reviewed-by: Dinan Gunawardena <dgunawardena@netronome.com>