main: move the SR-IOV code to a common location
The SR-IOV code is copied between nfp_main.c and nfp_net_main.c and
is identical save for a race which is present in only nfp_main.c.
Move it to a shared file.

This is a transitional change, we need the temporary new file because
for now there can only be a single *_main.c built into each module
(they all define module_init and module_exit functions).  Since this
new file will be removed in few commits no need to pay much attention
to its name.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>