nfp_net_compat: add missing includes
Compat code may redefine constants if the appropriate header
is not included before it.  We used to try to avoid adding
includes to compat code to minimise the risk of it masking
includes missing in real code.  It recently backfired in
RHEL 7.3 Beta so let's change our strategy and add all
relevant includes in compat code.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>