nfpcore: remove parent device pointer from nfp_cpp_ops
Struct nfp_cpp_operations contains a pointer to the parent
(PCI or platform) device.  It seems to be never needed in
low-level code and struct nfp_cpp contains a full struct
device (which include a parent pointer).  Remove the pointer
from ops structure and pass it to init function as parameter.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>