nfp_dev_cpp: interpret copy_{from|to}_user return codes correctly
copy_{from|to}_user() return number of bytes *not* copied.
The usual way of handling errors is to return -EFAULT when
copy_{from|to}_user() was not able to process all the data.
Make sure our code follows works this way.

Note that we will now return a negative code instead continuing
or returning a small positive integer in case of errors.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>
Acked-by: Francois H. Theron <francois.theron@netronome.com>