nfp: use NSP for FW loading
Instead of in-kernel CA replay code use the NSP to load firmware.
We no longer need to hold the device lock (NSP will grab it itself).
Since we are no longer using CA files change the FW file name
extensions from .cat to .nffw.

This is part of upstreaming effort, the CA replay code will not be
upstreamed.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>