nfp_pcie: don't use device managed version of ioremap
BAR mappings come and go.  The lifetime of an mapping is not the same
as the lifetime of the driver so it makes little sense to use devm_
functions other than avoiding leaks on unload.

Given that devm_ functions are not particularly liked on Linux netdev
mailing list, stop using them.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>