nfp_nsp: don't hold the lock entire time nfp_device is opened
Right now the NSP resource will be locked from the moment the first
command is issued until the nfp_device handle is closed.  This is
excessing and makes it impossible to communicate with the NSP from
user space.  Allow NSP to take control over the resource lock
and only lock it when command is being performed.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Acked-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>
Reviewed-by: david brunecz <david.brunecz@netronome.com>