Makefile: fix setting the NFP_SRC_VERSION
NFP_SRC_VERSION doesn't ever get set to git hash because
the nfp_src_ver variable is set on command line and then
not OVERRIDEn.  Move most of the code to Kbuild and use
export instead of command line to pass the contents of
.revision file if present.

NFP_SRC_VERSION seem to have never worked.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>