nfp: make comments networky
Linux net/ and drivers/net/ should use following comment style:

/* This is block
 * comment.
 */

and not:

/*
 * This is block
 * comment.
 */

In my experience it's OK to keep comments which are supposed
to separate sections of the file or are not really code related
(like licenses and other file headers) in the second format.
Make sure we are not stretching this rule.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>