nfp_nbi: Fix snprintf format specifier typo
The pause_poll_tx_flush_flags rtsym lookup is done with "i%2x" instead
of "i%d", which is the only legal island ID prefix format.  Because of
the '2' in format it also searches for "i 1.pause_poll_tx_flush_flags"
with a space in the name if the number is small.

Signed-off-by: Francois H. Theron <francois.theron@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>