nfp: limit message verbosity
We print quite a few lines on driver load.  Improve that by
not printing the cache BAR count and putting the serial number
in the same message as model and interface (note that we don't
have to worry if serial is null, printk takes care of that).

Don't print the "waiting for NSP to respond" message during NSP
noop wait, it could make users think that there is something
wrong with the driver<->NSP communication.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>