tests: fix ping test on multi port devices
Right now we only accept configuration for a single port, this
is obviously broken for multi port cards.  Make the addresses
and interface names into lists, and take configuration of multiple
of them.

Unlike FW projects we are not running serious traffic during the
tests so testing all ports in one go seems reasonable.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>