nfp: fix fall back to user space access only with PXE FW
After recent changes when PXE FW is loaded we will notice that
there is already a symbol table but it doesn't have PF control
memory symbol in it.  We will treat this situation as invalid FW
and not fall back to only enabling the user space access.  This
is a regression, make sure we do fall back there.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>