compat: protect from new kernels without switchdev enabled
The driver build would fail if we encountered a new (Linux 4.13+)
kernel without CONFIG_NET_SWITCHDEV enabled.  Make sure we don't
try to build flower app in that case and that we don't set the
switchdev ops.

We could try to degrade the flower feature set to not include
redirects (which is all for which we need switchdev at this point)
but it seems that this could lead to confusion and we also want to
avoid sprinkling the code with ifdefs as much as possible.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Simon Horman <simon.horman@netronome.com>