tests: add reading symbols from NFFW files
Add helper for reading symbols from NFFW files.  This will be useful
for upstream testing where we may want to inspect static symbols
to learn the FW capabilities, and the symbols are constant so we
don't really need to read them from the NFP we can just look into
the .data sections of the NFFW file.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>