tests: fix the tools setup test
The tools test is not supposed to load the module (it doesn't
need to), but this means the nfp id will be still set to None
when it's probing for the BSP tools.

Having the None as nfp id is a precaution, to make sure the id
is always refreshed on normal paths.  In this special case we
can safely hardwire the id (NFP is not actually accessed).

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>