tests: unbind driver from other devices
If there is more than one NFP in the system sometime this will
have adverse effects on the tests.  Random netdevs may appear
for instance, if the other cards are already pre-loaded with
firmware.  To prevent this from happening unbind devices which
are not the NFP under test from the driver after each load.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>