tests: bpf: add atomic add tests
Add tests for atomic add.  Check we reject programs which try to
read the counters, or do xadd not on map value.  Make sure the
values are added, byteswapped and overflow correctly.  Make sure
kernel rejects unaligned accesses.  Check the three variants of
codegen - both large and 16bit or only one of them.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Jiong Wang <jiong.wang@netronome.com>