tools: fix IRQ affinity xps_mask for machines with more than 32 cores
xps_mask is formatted as an array of 32 bit fields, in hex, separated
by commas.  The comma is currently missed.  Use sed to insert comma
between the 63..32 and 31..0 CPU masks.  This is not sufficient for
machines with more than 64 CPUs, but other aspects of the script
also do not support more than 64 CPUs currently (e.g. where << is
used).

Signed-off-by: Michael Rapson <michael.rapson@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>