tests: rework reboot tests
Since we now have the ability to rerun MAC init without requiring a
system reboot, we can modify the tests that required a reboot to only
require a driver reload.

NOTE: The set_speed test can now be executed with a Flower firmware image,
using representors. The split_port test is not supported by the Flower
app yet.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>