compat: migrate NTH to new DebugFS file locking
From Linux 4.15 on locking in DebugFS no longer depends on SRCU.
Old API has been removed, move to the new one.

While at it move the wrapper out of kcompat code, we are ignoring
locking completely on kernels older than 4.7, which is probably
OK in the test code, but we shouldn't do that in real code.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Simon Horman <simon.horman@netronome.com>