tests: bpf: switch to Clang for compiling .S file
It is better to use Clang instead of llvm-mc to compile .S file.  We could
then utilize the preprocessor to use the same macros defined in C header
files. This could keeps some magic constants in consistent with those in
C tests.

Also, we'd prefer to use "//" as comment prefix instead of "#" after this
switch. The latter has special meaning for preprocessor, for example, "#if"
etc.

Signed-off-by: Jiong Wang <jiong.wang@netronome.com>