tools: add script to extract firmware MIP name
This script searches the current default locations where NFP firmware
files generally exists and prints the MIP name/version for each of the
firmware
images found there.

Alternatively, as an argument one can also specify a directory to search
or a file to extract the firmware MIP name/version from.

The output of the script is in this format:
<image filename>  <MIP name/version>  [symlink]
...

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>