tests: bpf: support eBPF JIT code-gen scan
For JIT optimisation, especially instructions transformation, runtime tests
themselves are not enough as we want to make sure the optimisation and the
related transformations had really happened. Therefore, we need an test
infrastructure which allow us to scan the code-gen of JIT compiler.

This patch implements this feature. And the following check rules are
supported:

The specified pattern exists, no matter of times.
===
  /* CHECK-CODEGEN: .*local_csr_rd */

The specified pattern exists, and shown up for N times.
===
  /* CHECK-CODEGEN-TIMES-N: .*local_csr_rd */

The specified pattern does *not* exist.
===
  /* CHECK-CODEGEN-NOT: .*mem\[write32_swap.*8\].* */

The match rules could just be embedded as comments in test source, they are
also free to be put in an separate file, the method bpf_jit_patterns_file
should be overiden to specify the file name containing the patterns.

The patterns are using the same regexp syntax as python RE library.

This feature is supported on all eBPF testcases, both those in samples/bpf
and those in samples/xdp.

NOTE: The match happens on line basis.

Signed-off-by: Jiong Wang <jiong.wang@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>