tests: bpf: fix checks for eBPF objects number in bpf_refcnt test
Commit 4908f162612c ("tests: bpf: don't depend on being the only BPF
user") moved object counting in eBPFrefcnt test class from
has_bpf_objects() to bpf_obj_cnt(). The method has_bpf_objects() now
relies on bpf_obj_cnt() returning a value different from the one
observed at instance creation, which means that it does not tell whether
there is any eBPF object loaded on the system, but rather if the number
of existing objects has changed between the creation of this test
instance and the call to the method.

As a consequence, has_bpf_objects() can no more be used to make sure
that no objects are present at instance creation time, as it would only
check that the initial number is equal to itself. We could fix it by
using bpf_obj_cnt() instead. However, this would defeat the purpose of
the patch that brought this change. Therefore, we simply suppress the
check for existing eBPF objects prior to running the test.

Furthermore, rename has_bpf_objects() into something that better
reflects what the function verifies now (bpf_objects_nb_differs()). Also
update the Exception() message, to remain consistent with what we check.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>