tests: bpf: add basic extack and verifier log tests for TC offload
Add functions to check for extack messages to class CommonTest, and use
them in tc_bpf_load() function (called in particular by the prepare()
method) when the relevant arguments are provided.

"extack" is used to validate that this "extack" message appears in
stderr when the filter is attached. "needle_noextack" is a string that
must not appear in stderr. Finally, the presence of the string in
"verifier_log" is checked in the buffer from the kernel verifier, also
printed to stderr.

All tests that produce extack or verifier output for eBPF offload with
TC are adapted to perform the related checks. No new tests are added.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>