tests: bpf: add test for spurious extack messages
Add a new (TC) eBPF test in order to try to catch leaking extack
messages. This test is inspired by kernel selftests offload tests, run
on netdevsim module. This test search the output of "tc filter" commands
for spurious extack messages, for several operations (add, replace,
delete), with and without hardware offload required on the command line,
with and without hardware offload enabled on the card.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>