tests: bpf: extend bpf_mtu_check test for XDP
Extend test for MTU verification, which was only run with TC offload so
far. Check 1) that XDP program cannot be offloaded when MTU is too high,
and 2) that MTU cannot be changed to a high value when a XDP program is
running on the NIC.

For the first case, also use the recently added check_extack method to
verify that the expected extack message is printed.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>