tests: unit: update invalid fw dump levels
Recently, CoreNIC has also implemented the firmware dump feature. Since
both CoreNIC and Disa firmware supports this, the bad ethtool dump level
check has become obsolete and can be removed.

The flag is not validated and all dump levels will be accepted. This is
arguably something that needs to be fixed, but since there are proposals
for updates to this feature in any case, we can handle it at that point.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>