tests: define reprs and phys_netdevs
Some tests want to know which netdevs are usable for operating on the
physical ports of the NFP. This differs between firmware/apps that use
vNICs and ones that use representors.

So we define the reprs and phys_netdev lists that can now be used for
this purpose.

While at it we update the reload tests to make use of these lists.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>