tests: don't unbind higher order PCIe units
The test infrastructure generally unbinds all NFP devices from the
driver that doesn't match the supplied serial number.

However, when using devices with multiple PCIe interfaces, they share a
serial number. In this case, we actually want the interface #0 stored in
the test pci_id and pci_dbdf, and additionally store the set of pci_ids
and pci_dbdfs. If we don't, we could end up with the tests being run on
non-zero PCIe interface which with firmware unloaded already by the PCIe

We only unbind devices that aren't part of that list, thus we will have
all PCIe devices bound for a test run.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>