tests: support testing pre-installed drivers
Swap out the 'upstream_drv' config option for 'installed_drv'.
If set, we modprobe the modules (incl. nfp_test_harness.ko if
available).

If the version of the driver doesn't include nfp_dev_cpp module
parameter, we assume this is an upstream test and as such don't
attempt to load the nfp_test_harness module.

To install the nfp_test_harness module, use the following command:
 $ make CONFIG_NFP=n CONFIG_NFP_TEST_HARNESS=m install

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>