tests: unit: skip eth table for DACs
DACs don't support enable/disable since this is a module capability
which isn't generally available for DAC modules.

After doing the initial comparison, we skip the remainder of the test
for all interfaces which register a copper type. If all ports are
skipped, we skip the entire test from that point.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>