tools: extract version and buildnumber from elf info
Recent toolchain updates added version and build numbers to the
build_info portion of the elf files.
When available, we read those and print them too.

In order to prevent breakage of existing scripts relying on the
ordering, we add the version and buildnumber to the name, separated by
'/' characters.

Signed-off-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>